chore: refactor methods into react components
